+++ vivisect/parsers/vpdb.py
@@ -1,9 +1,11 @@
-import vstruct.defs.pepdb as v_d_pdb
+import vstruct.defs.vpdb as v_d_pdb
+
+from vstruct.primitives import v_uint32
 
 class PDB:
     def __init__(self, byts):
         self.byts = byts
-        self.superblock = SuperBlock()
+        self.superblock = v_d_pdb.SuperBlock()
         self.superblock.vsParse(byts)
 
         self.streamsdirs = self.getStreamDirs()
@@ -13,7 +15,7 @@ class PDB:
         The stream directory is not guaranteed to fit in a single block
         so an MSF adds a layer of indirection.
         '''
-        block = self.getBlockByIdx(self.superblock.blockmapaddr)
+        block = self.getBlockByIdx(self.superblock.blockMapAddr)
         if not block:
             return []
         valus = []
@@ -30,7 +32,7 @@ class PDB:
         dirs = []
         for idx in valus:
             block = self.getBlockByIdx(idx)
-            sdir = StreamDirectory(int(self.superblock.blocksize))
+            sdir = v_d_pdb.StreamDirectory(int(self.superblock.blockSize))
             sdir.vsParse(block)
 
             dirs.append(sdir)
@@ -38,7 +40,7 @@ class PDB:
         return dirs
 
     def getBlockByIdx(self, idx):
-        bsize = self.superblock.blocksize
+        bsize = self.superblock.blockSize
         offset = idx * bsize
         if offset < len(self.byts):
             return self.byts[offset:offset+bsize]
@@ -48,13 +50,13 @@ class PDB:
         idx = 0
         offset = 0
         while offset < len(self.byts):
-            yield idx, self.byts[offset:(offset+self.superblock.blocksize)]
+            yield idx, self.byts[offset:(offset+self.superblock.blockSize)]
             idx += 1
-            offset += self.superblock.blocksize
+            offset += self.superblock.blockSize
 
 if __name__ == '__main__':
-    with open('C:\\Users\\James\\Documents\\Visual Studio 2015\\Projects\\devicelist\\Debug\\devicelist.pdb', 'rb') as fd:
+    with open('C:\\Users\\James\\Documents\\Visual Studio 2015\\Projects\\raytracing\\Debug\\raytracing.pdb', 'rb') as fd:
         byts = fd.read()
-    pdb = PDB(byts)
+    obj = PDB(byts)
     breakpoint()
     print('wat')
